move transaction utility into its own file
